+++ test/commit_table_builder_test.go
@@ -3,15 +3,14 @@ package test
 import (
 	"testing"
 
-	"github.com/Originate/git-town/test/gherkintools"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestCommitTableBuilder(t *testing.T) {
 	builder := NewCommitTableBuilder()
-	commit1 := gherkintools.Commit{SHA: "sha1", Branch: "branch1", Message: "commit1"}
-	commit2 := gherkintools.Commit{SHA: "sha2", Branch: "branch2", Message: "commit2"}
-	commit3 := gherkintools.Commit{SHA: "sha3", Branch: "branch3", Message: "commit3"}
+	commit1 := Commit{SHA: "sha1", Branch: "branch1", Message: "commit1"}
+	commit2 := Commit{SHA: "sha2", Branch: "branch2", Message: "commit2"}
+	commit3 := Commit{SHA: "sha3", Branch: "branch3", Message: "commit3"}
 	builder.Add(commit1, "local")
 	builder.Add(commit1, "remote")
 	builder.Add(commit2, "local")
